requirements.txt
streamlit==1.56.0
python-dotenv==1.2.2
langchain==1.2.15
langchain-core==1.2.27
langchain-community==0.4.1
langchain-huggingface==1.2.1
langchain-google-genai==4.2.1
langchain-text-splitters==1.1.1
faiss-cpu==1.13.2
sentence-transformers==5.3.0
pypdf==6.9.2
google-generativeai==0.8.6
